requirements.txt

fastapi
tensorflow-cpu
uvicorn
Pillow
python-multipart
protobuf
